Server/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==1.4.46
alembic==1.13.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
loguru==0.7.2
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.26.0
psutil==6.1.0
websockets==13.1
pytest==7.4.3
pytest-asyncio==0.21.1
cryptography==41.0.7
psycopg2-binary==2.9.9
prisma==0.13.0